02-llm-finetuning-alignment/src/requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
banal==1.0.6
boto3==1.38.32
botocore==1.38.32
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
dataset==1.6.2
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
docker==7.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mock==4.0.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.18
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathos==0.3.4
pexpect==4.9.0
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
sagemaker==2.246.0
sagemaker-core==1.0.37
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
SQLAlchemy==1.4.54
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tblib==3.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trl==0.18.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
python-dotenv